dashboard/requirements.txt
alembic==1.4.3
astroid==2.6.6
bidict==0.21.4
cachelib==0.4.1
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==1.0.2
Flask-Admin==1.5.4
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-MySQL==1.4.0
Flask-Session==0.3.2
Flask-SocketIO==5.1.1
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.0.4
importlib-metadata==4.8.3
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.6.0
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
pycparser==2.21
pylint==2.8.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMySQL==0.9.3
pyodbc==4.0.32
python-dateutil==2.8.0
python-dotenv==0.17.1
python-editor==1.0.4
python-engineio==4.3.0
python-socketio==5.4.1
six==1.12.0
SQLAlchemy==1.2.17
toml==0.10.2
typed-ast==1.4.3
typing_extensions==4.1.1
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.2.1
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
